♻️Refactor: Separation signup component container
